+++ CheckInManager.sln
@@ -5,21 +5,23 @@ VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{68E961A5-A765-4024-8474-ED5F27A00B93}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CheckInManager.CupsPrinter", "src\CheckInManager.CupsPrinter\CheckInManager.CupsPrinter.csproj", "{ACAD0D4A-EE54-4456-A0AA-FC7E32571741}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CheckInManager.Core", "src\CheckInManager.Core\CheckInManager.Core.csproj", "{EF0D803C-BF03-4BD1-AC75-1C37245CC940}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CheckInManager.Api.QRGenerator", "src\CheckInManager.Api.QRGenerator\CheckInManager.Api.QRGenerator.csproj", "{BFA491D4-6CDA-4499-816C-9A41A46DE812}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CheckInManager.QRGenerator", "src\CheckInManager.QRGenerator\CheckInManager.QRGenerator.csproj", "{3D1FE9D3-7768-4EDA-9DF2-77B63EBD1FCA}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CheckInManager.Api.Printer", "src\CheckInManager.Api.Printer\CheckInManager.Api.Printer.csproj", "{1ADAA309-8ECA-452B-9E16-78859B21CE15}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CheckInManager.CupsPrinter", "src\CheckInManager.CupsPrinter\CheckInManager.CupsPrinter.csproj", "{ACAD0D4A-EE54-4456-A0AA-FC7E32571741}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CheckInManager.Core", "src\CheckInManager.Core\CheckInManager.Core.csproj", "{EF0D803C-BF03-4BD1-AC75-1C37245CC940}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CheckInManager.Api.QRGenerator", "src\CheckInManager.Api.QRGenerator\CheckInManager.Api.QRGenerator.csproj", "{BFA491D4-6CDA-4499-816C-9A41A46DE812}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CheckInManager.QRGenerator", "src\CheckInManager.QRGenerator\CheckInManager.QRGenerator.csproj", "{3D1FE9D3-7768-4EDA-9DF2-77B63EBD1FCA}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CheckInManager.Api.Printer", "src\CheckInManager.Api.Printer\CheckInManager.Api.Printer.csproj", "{1ADAA309-8ECA-452B-9E16-78859B21CE15}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{07601C33-801F-4595-A294-3D331D0BABBF}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{07601C33-801F-4595-A294-3D331D0BABBF}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CheckInManager.CupsPrinter.Tests", "tests\CheckInManager.CupsPrinter.Tests\CheckInManager.CupsPrinter.Tests.csproj", "{EF979182-28B7-41E0-A0B0-31BBB3CD46A8}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CheckInManager.Core.Tests", "test\CheckInManager.Core.Tests\CheckInManager.Core.Tests.csproj", "{42D105A2-5294-4E11-9A25-8BB74FEED6DE}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CheckInManager.QRGenerator.Tests", "tests\CheckInManager.QRGenerator.Tests\CheckInManager.QRGenerator.Tests.csproj", "{D1298CE7-AA0E-4CAD-A563-F493F0CEA808}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CheckInManager.QRGenerator.Tests", "test\CheckInManager.QRGenerator.Tests\CheckInManager.QRGenerator.Tests.csproj", "{D1298CE7-AA0E-4CAD-A563-F493F0CEA808}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CheckInManager.CupsPrinter.Tests", "test\CheckInManager.CupsPrinter.Tests\CheckInManager.CupsPrinter.Tests.csproj", "{EF979182-28B7-41E0-A0B0-31BBB3CD46A8}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,6 +29,14 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{EF0D803C-BF03-4BD1-AC75-1C37245CC940}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EF0D803C-BF03-4BD1-AC75-1C37245CC940}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EF0D803C-BF03-4BD1-AC75-1C37245CC940}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EF0D803C-BF03-4BD1-AC75-1C37245CC940}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3D1FE9D3-7768-4EDA-9DF2-77B63EBD1FCA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3D1FE9D3-7768-4EDA-9DF2-77B63EBD1FCA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3D1FE9D3-7768-4EDA-9DF2-77B63EBD1FCA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3D1FE9D3-7768-4EDA-9DF2-77B63EBD1FCA}.Release|Any CPU.Build.0 = Release|Any CPU
 		{ACAD0D4A-EE54-4456-A0AA-FC7E32571741}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{ACAD0D4A-EE54-4456-A0AA-FC7E32571741}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{ACAD0D4A-EE54-4456-A0AA-FC7E32571741}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -39,33 +49,30 @@ Global
 		{1ADAA309-8ECA-452B-9E16-78859B21CE15}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1ADAA309-8ECA-452B-9E16-78859B21CE15}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1ADAA309-8ECA-452B-9E16-78859B21CE15}.Release|Any CPU.Build.0 = Release|Any CPU
-		{EF0D803C-BF03-4BD1-AC75-1C37245CC940}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EF0D803C-BF03-4BD1-AC75-1C37245CC940}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EF0D803C-BF03-4BD1-AC75-1C37245CC940}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EF0D803C-BF03-4BD1-AC75-1C37245CC940}.Release|Any CPU.Build.0 = Release|Any CPU
-		{3D1FE9D3-7768-4EDA-9DF2-77B63EBD1FCA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3D1FE9D3-7768-4EDA-9DF2-77B63EBD1FCA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3D1FE9D3-7768-4EDA-9DF2-77B63EBD1FCA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3D1FE9D3-7768-4EDA-9DF2-77B63EBD1FCA}.Release|Any CPU.Build.0 = Release|Any CPU
-		{EF979182-28B7-41E0-A0B0-31BBB3CD46A8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EF979182-28B7-41E0-A0B0-31BBB3CD46A8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EF979182-28B7-41E0-A0B0-31BBB3CD46A8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EF979182-28B7-41E0-A0B0-31BBB3CD46A8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{42D105A2-5294-4E11-9A25-8BB74FEED6DE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{42D105A2-5294-4E11-9A25-8BB74FEED6DE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{42D105A2-5294-4E11-9A25-8BB74FEED6DE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{42D105A2-5294-4E11-9A25-8BB74FEED6DE}.Release|Any CPU.Build.0 = Release|Any CPU
 		{D1298CE7-AA0E-4CAD-A563-F493F0CEA808}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{D1298CE7-AA0E-4CAD-A563-F493F0CEA808}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D1298CE7-AA0E-4CAD-A563-F493F0CEA808}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D1298CE7-AA0E-4CAD-A563-F493F0CEA808}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EF979182-28B7-41E0-A0B0-31BBB3CD46A8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EF979182-28B7-41E0-A0B0-31BBB3CD46A8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EF979182-28B7-41E0-A0B0-31BBB3CD46A8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EF979182-28B7-41E0-A0B0-31BBB3CD46A8}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
+		{EF0D803C-BF03-4BD1-AC75-1C37245CC940} = {68E961A5-A765-4024-8474-ED5F27A00B93}
+		{3D1FE9D3-7768-4EDA-9DF2-77B63EBD1FCA} = {68E961A5-A765-4024-8474-ED5F27A00B93}
 		{ACAD0D4A-EE54-4456-A0AA-FC7E32571741} = {68E961A5-A765-4024-8474-ED5F27A00B93}
 		{BFA491D4-6CDA-4499-816C-9A41A46DE812} = {68E961A5-A765-4024-8474-ED5F27A00B93}
 		{1ADAA309-8ECA-452B-9E16-78859B21CE15} = {68E961A5-A765-4024-8474-ED5F27A00B93}
-		{EF0D803C-BF03-4BD1-AC75-1C37245CC940} = {68E961A5-A765-4024-8474-ED5F27A00B93}
-		{3D1FE9D3-7768-4EDA-9DF2-77B63EBD1FCA} = {68E961A5-A765-4024-8474-ED5F27A00B93}
-		{EF979182-28B7-41E0-A0B0-31BBB3CD46A8} = {07601C33-801F-4595-A294-3D331D0BABBF}
+		{42D105A2-5294-4E11-9A25-8BB74FEED6DE} = {07601C33-801F-4595-A294-3D331D0BABBF}
 		{D1298CE7-AA0E-4CAD-A563-F493F0CEA808} = {07601C33-801F-4595-A294-3D331D0BABBF}
+		{EF979182-28B7-41E0-A0B0-31BBB3CD46A8} = {07601C33-801F-4595-A294-3D331D0BABBF}
 	EndGlobalSection
 EndGlobal
