+++ Directory.Build.props
@@ -1,9 +1,11 @@
 <Project>
     <PropertyGroup>
         <TargetFramework>net8.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
         <LangVersion>latest</LangVersion>
+
+        <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
+
         <ArtifactsPath>$(MSBuildThisFileDirectory)artifacts</ArtifactsPath>
     </PropertyGroup>
 </Project>
