+++ drivers/cups/nemonic/nemonic_mini.icns
