+++ src/CheckInManager.Core/CheckInManager.Core.csproj
@@ -3,8 +3,9 @@
   <PropertyGroup>
     <!-- <TargetFramework>netstandard2.0</TargetFramework> -->
     <OutputType>Library</OutputType>
-    <RootNamespace>CheckInManager.Core</RootNamespace>
+
     <AssemblyName>CheckInManager.Core</AssemblyName>
+    <RootNamespace>CheckInManager.Core</RootNamespace>
   </PropertyGroup>
 
 </Project>
