+++ src/CheckInManager.CupsPrinter/CheckInManager.CupsPrinter.csproj
@@ -3,8 +3,9 @@
   <PropertyGroup>
     <!-- <TargetFramework>netstandard2.0</TargetFramework> -->
     <OutputType>Library</OutputType>
-    <RootNamespace>CheckInManager.CupsPrinter</RootNamespace>
+
     <AssemblyName>CheckInManager.CupsPrinter</AssemblyName>
+    <RootNamespace>CheckInManager.CupsPrinter</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
