+++ src/CheckInManager.QRGenerator/CheckInManager.QRGenerator.csproj
@@ -3,8 +3,9 @@
   <PropertyGroup>
     <!-- <TargetFramework>netstandard2.0</TargetFramework> -->
     <OutputType>Library</OutputType>
-    <RootNamespace>CheckInManager.QRGenerator</RootNamespace>
+
     <AssemblyName>CheckInManager.QRGenerator</AssemblyName>
+    <RootNamespace>CheckInManager.QRGenerator</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
